chore: update build tools

+++ build/build.sh
@@ -2,7 +2,7 @@
 # @Author: Just be free
 # @Date:   2021-06-29 13:45:56
 # @Last Modified by:   Just be free
-# @Last Modified time: 2021-06-29 15:53:07
+# @Last Modified time: 2021-09-01 11:02:13
 
 
 function build {
@@ -10,8 +10,23 @@ function build {
   npm run build
   echo "build success"
   cd ../..
-  rm -rf docs/$1
-  mkdir docs/$1
+  if [ ! -d "docs/" ];then
+    mkdir docs
+    mkdir docs/$1
+    # if [ ! -d "docs/$1" ];then
+  else
+    if [ ! -d "docs/$1" ];then
+      # echo "文件夹不存在"
+      mkdir docs/$1
+    else
+      rm -rf docs/$1
+      mkdir docs/$1
+      # echo "$1 folder exists, skip this message"
+    fi
+    # rm -rf docs/$1
+  fi
+  # rm -rf docs/$1
+  # mkdir docs/$1
   mv packages/$1/docs/* docs/$1
 }
